info.xlsx移动道Resource并改为嵌入的资源, 默认加载程序同目录的文件, 没有再加载嵌入的资源; 宝箱隐藏道具界面添加完成进度显示功能;  调整弱点和宝箱隐藏道具界面展示方式, 添加筛选及总进度显示功能; 更换一个程序图标;

+++ OctopathTraveler/SaveData.cs
@@ -6,6 +6,12 @@ namespace OctopathTraveler
 {
     class SaveData
     {
+#if DEBUG_READONLYMODE
+        public static readonly bool IsReadonlyMode = true;
+#else
+        public static readonly bool IsReadonlyMode = false;
+#endif
+
         private static SaveData mThis;
         private string mFileName = null;
         private byte[] mBuffer = null;
@@ -32,14 +38,14 @@ namespace OctopathTraveler
 
         public bool Save()
         {
-            if (mFileName == null || mBuffer == null) return false;
+            if (IsReadonlyMode || mFileName == null || mBuffer == null) return false;
             File.WriteAllBytes(mFileName, mBuffer);
             return true;
         }
 
         public bool SaveAs(string filenname)
         {
-            if (mBuffer == null) return false;
+            if (IsReadonlyMode || mBuffer == null) return false;
             mFileName = filenname;
             return Save();
         }
@@ -191,6 +197,9 @@ namespace OctopathTraveler
 
         private void Backup()
         {
+            if (IsReadonlyMode)
+                return;
+
             DateTime now = DateTime.Now;
             string path = Path.Combine(Directory.GetCurrentDirectory(), "backup");
             if (!Directory.Exists(path))
